:speaking_head: Convert Lexer to Stream Transform
**Breaking:** This is known to break the build. But to make these
commits of manageable size, it's going to make most sense to do this in
pieces.

This converts the Lexer into a Transform stream processor. This is in an
effort to make the lexer and parser more independent from one another.

This is necessary because some lexed content will come in over multiple
lines, and also may come in separate chunks. Combined with the
`NewlineTransform`, it should result in a more predictable set of lexed
content for the parser to be able to parse.