:construction_worker: Convert Parser to Transform
**Breaking:** This is known to break the build. But to make these
commits of manageable size, it's going to make most sense to do this in
pieces.

The last piece of this is a **major** rewrite of the existing parser to
make it a Transform. This removes the old functionality completely and
switches over to the new parsing/lexing methods.

Ya gotta break some eggs to make an omelette :fried_egg: