:dog2: Add UID/Date Parser for Fetch
Adding the first pieces of parsing for the Fetch command. This is one of
the most complicated pieces of the spec, so this parsing will be added
over many different commits. But this commit lays much of the groundwork
for things like integration tests and general code structure.

Test plan:
yarn jest