:palms_up_together: Base & Noop Command Classes
This is the beginning of the new command style strcture. Commands will
be added as self managing classes that know how to generate and send the
actual command text as well as process parsed responses.

The base class implements a common strategy of sending a tagged command
and listening for a tagged response, capturing everything in between.

The no-op is the most simple implementation of this command system.