:runner: Create basic IMAPSession Class
This class will eventually serve as the base class that is exposed and
that other systems would interface with. This will contain the public-
facing API for the library and will be both a mirror and an update of
existing functionality.

Note that this commit does not expose this class nor remove or change
the original `Connection` class which currently acts as the public
interface. Eventually the `Connection` class will be cleaned up and it
will serve purely as the connection for the session without any logic
for actual commands.