😌 Rewrite Connection Class
Pulled everything out of the connection class that is not actually
related to the connection. This means almost all IMAP related commands
and controls will be in a new class. The code went from almost 2k lines
to around 300, which should be a lot easier to understand.

From a practical perspective, I chose to write a brand new class and
just remove the old copy completely. The old code was inter-weaved with
a lot of implicit assumptions and untangling that would have been a
nightmare. This does, however, mean this is all (or almost all) new
code. Which is extra terrifying. But that's a future me problem.